Split isFresh into a trait & add config timestamp checking.
To reduce code duplication move isFresh() into a trait. Also add checks
for config timestamp vs. the target mtime. This lets config changes
invalidate the builds which is useful when adding 'old' files to
a target.
